Fix log collection of virtv2v pods with long names
Virtv2v pods are named based on the plan name and VM id. If
the plan and VM id are too long, the name of the pod will be
truncated and not contain the entire plan or VM id. Targeted
log collection will fail for such a pod, as the collection
script attempts to grep for the entire VM id in the pod name.

Because the collection script already filters for pods based
on the vmID label, this additional grep for the VM id in the
pod name should be redundant. Removing it allows the pod to be
found and logs collected.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=2064336